src/mcp/requirements.txt
# MCP Module Requirements
# =======================

# Core dependencies
pydantic>=2.0.0
aiohttp>=3.8.0

# Optional dependencies for full functionality
PyYAML>=6.0  # For YAML configuration support
python-jose>=3.3.0  # For JWT token handling
cryptography>=41.0.0  # For secure operations

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0